# documentation/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "AppController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "DB connectivity OK"
          }
        },
        "summary": "Healthcheck",
        "tags": [
          "system"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Login and get a JWT access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User info from JWT payload"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/files": {
      "get": {
        "operationId": "FileController_listFiles",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtre sur le statut des fichiers",
            "schema": {
              "default": "all",
              "example": "active",
              "type": "string",
              "enum": [
                "all",
                "active",
                "expired",
                "deleted"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Numéro de page (>= 1)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Taille de page (1..100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List files for current user",
        "tags": [
          "files"
        ]
      }
    },
    "/api/files/{id}": {
      "delete": {
        "operationId": "FileController_deleteFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user file (soft delete + best-effort disk delete)",
        "tags": [
          "files"
        ]
      }
    },
    "/api/files/upload": {
      "post": {
        "operationId": "FileController_uploadFile",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload a file (multipart/form-data)",
        "tags": [
          "files"
        ]
      }
    },
    "/download/{token}/meta": {
      "get": {
        "operationId": "DownloadController_getMeta",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Download token (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get public file metadata by token",
        "tags": [
          "public-download"
        ]
      }
    },
    "/download/{token}": {
      "get": {
        "operationId": "DownloadController_downloadUnprotected",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Download token (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Download file by token (unprotected or via password POST)",
        "tags": [
          "public-download"
        ]
      },
      "post": {
        "operationId": "DownloadController_downloadProtected",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Download token (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Download protected file by token + password",
        "tags": [
          "public-download"
        ]
      }
    },
    "/api/download/{token}/meta": {
      "get": {
        "operationId": "ApiDownloadController_getMeta",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Download token (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get download metadata (API variant)",
        "tags": [
          "download"
        ]
      }
    },
    "/api/download/{token}": {
      "get": {
        "operationId": "ApiDownloadController_downloadUnprotected",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Download token (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Download file (API variant)",
        "tags": [
          "download"
        ]
      },
      "post": {
        "operationId": "ApiDownloadController_downloadProtected",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Download token (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Download protected file with password (API variant)",
        "tags": [
          "download"
        ]
      }
    }
  },
  "info": {
    "title": "DataShare API",
    "description": "API du projet DataShare (auth + fichiers + téléchargement public).",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Adresse email de l'utilisateur",
            "example": "alice@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Mot de passe (min 8 caractères)",
            "example": "P4ssw0rd!",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Adresse email de l'utilisateur",
            "example": "alice@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Mot de passe",
            "example": "P4ssw0rd!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "DownloadPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Mot de passe du fichier (min 6 caractères)",
            "example": "secret1",
            "minLength": 6
          }
        },
        "required": [
          "password"
        ]
      }
    }
  }
}